backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
groq==0.4.1
chromadb==0.4.22
requests==2.31.0
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
firebase-admin==6.4.0
pydantic==2.5.3
